feat: use Long for ids

+++ teaching-center-backend/src/main/java/com/breakingbitstudio/teaching_center_backend/dto/request/CreateClassroomRequest.java
@@ -12,7 +12,7 @@ public record CreateClassroomRequest(
 
         String thumbnailUrl,
 
-        Integer teacherId,
+        Long teacherId,
 
         @Positive
         Integer totalSessions,
