+++ teaching-center-backend/src/main/java/com/breakingbitstudio/teaching_center_backend/dto/request/PatchClassroomRequest.java
@@ -12,7 +12,7 @@ public record PatchClassroomRequest(
 
         String thumbnailUrl,
 
-        Integer teacherId,
+        Long teacherId,
 
         @Positive
         Integer totalSessions,
