+++ teaching-center-backend/src/main/java/com/breakingbitstudio/teaching_center_backend/dto/request/UpdateClassroomRequest.java
@@ -14,7 +14,7 @@ public record UpdateClassroomRequest(
 
         String thumbnailUrl,
 
-        Integer teacherId,
+        Long teacherId,
 
         @Positive
         Integer totalSessions,
